requirements.txt
Django==1.11.16
djangorestframework==3.7.7
django-rest-auth==0.9.3
django-filter==2.0.0